Group the type alias cross-referencing test with Python domain tests
It is not, strictly speaking, an autodoc test.
