--- tests/roots/test-ext-autodoc-type-alias-xref/conf.py
@@ -1,11 +1,0 @@
-import sys
-from pathlib import Path
-
-sys.path.insert(0, str(Path.cwd().resolve()))
-
-extensions = ['sphinx.ext.autodoc']
-nitpicky = True
-autodoc_type_aliases = {
-    'pathlike': 'pathlike',
-    'Handler': 'Handler',
-}
