--- tests/roots/test-ext-autodoc-type-alias-xref/index.rst
@@ -1,7 +1,0 @@
-Type Alias Cross-Reference Test
-===============================
-
-This tests that type aliases in function signatures can be cross-referenced properly.
-
-.. automodule:: alias_module
-   :members:
