+++ tests/test_ext_autodoc/test_ext_autodoc.py
@@ -2607,6 +2607,21 @@ def test_autodoc_pep695_type_alias(app):
         '   This is class Foo.',
         '',
         '',
+        '.. py:data:: Handler',
+        '   :module: target.pep695',
+        '',
+        '   A generic type alias',
+        '',
+        '   alias of :py:class:`type`\\ [:py:class:`Exception`]',
+        '',
+        '',
+        '.. py:type:: HandlerTypeAliasType',
+        '   :module: target.pep695',
+        '   :canonical: type[Exception]',
+        '',
+        '   This is an explicitly constructed generic alias typing.TypeAlias.',
+        '',
+        '',
         '.. py:type:: Pep695Alias',
         '   :module: target.pep695',
         '   :canonical: ~target.pep695.Foo',
@@ -2647,6 +2662,43 @@ def test_autodoc_pep695_type_alias(app):
         '   :canonical: ~target.pep695.Foo',
         '',
         '   This is an explicitly constructed typing_extensions.TypeAlias.',
+        '',
+        '',
+        '.. py:function:: buffer_len(data: buffer_like) -> int',
+        '   :module: target.pep695',
+        '',
+        '   Return length of a buffer-like object.',
+        '   Tests Union type alias cross-reference resolution.',
+        '',
+        '',
+        '.. py:data:: buffer_like',
+        '   :module: target.pep695',
+        '   :value: bytes | bytearray | memoryview',
+        '',
+        '   Some buffer-like object',
+        '',
+        '',
+        '.. py:data:: pathlike',
+        '   :module: target.pep695',
+        '   :value: str | pathlib._local.Path',
+        '',
+        '   Any type of path',
+        '',
+        '',
+        '.. py:function:: process_error(handler: Handler, other: ~target.pep695.HandlerTypeAliasType) -> str',
+        '   :module: target.pep695',
+        '',
+        '   Process an error with a custom handler type.',
+        '',
+        '   Tests generic type alias cross-reference resolution.',
+        '',
+        '',
+        '.. py:function:: read_file(path: pathlike) -> bytes',
+        '   :module: target.pep695',
+        '',
+        '   Read a file and return its contents.',
+        '',
+        '   Tests Union type alias cross-reference resolution.',
         '',
         '',
         '.. py:function:: ret_pep695(a: ~target.pep695.Pep695Alias) -> ~target.pep695.Pep695Alias',
